# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Task:
        properties:
            assigned_user_id:
                format: int64
                type: integer
                x-go-name: AssignedUserID
            comments:
                items:
                    type: string
                type: array
                x-go-name: Comments
            completed:
                type: boolean
                x-go-name: Completed
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            description:
                type: string
                x-go-name: Description
            id:
                format: int64
                type: integer
                x-go-name: ID
            items:
                items:
                    type: string
                type: array
                x-go-name: Items
            title:
                type: string
                x-go-name: Title
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
        type: object
        x-go-package: tms.zinkworks.com/model
    TaskComment:
        properties:
            comment:
                type: string
                x-go-name: Comment
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            id:
                format: int64
                type: integer
                x-go-name: ID
            task_id:
                format: int64
                type: integer
                x-go-name: TaskID
        type: object
        x-go-package: tms.zinkworks.com/model
host: localhost:4000
info:
    contact:
        email: support@taskmanagementsystem.com
        name: Support Team
    description: |-
        This is a sample Task Management System server.
        It provides operations to manage tasks, assign users, and manage comments.
    license:
        name: MIT
        url: http://opensource.org/licenses/MIT
    title: Task Management System API.
    version: 1.0.0
paths:
    /comments:
        post:
            consumes:
                - application/json
            description: Inserts a new task comment into the database.
            operationId: createTaskCommentsEndpoint
            produces:
                - application/json
            responses:
                "201":
                    $ref: '#/responses/taskCommentCreatedResponse'
                "400":
                    $ref: '#/responses/badRequestError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Create a task comment.
            tags:
                - tasks
    /comments/{taskID}:
        get:
            description: Returns a list of comments for a task based on its ID.
            operationId: getTaskCommentsEndpoint
            parameters:
                - description: The ID of the task whose comments are to be retrieved.
                  format: int64
                  in: path
                  name: taskID
                  required: true
                  type: integer
                  x-go-name: TaskID
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/allCommentsResponse'
                "400":
                    $ref: '#/responses/invalidIdError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Get comments for a specific task.
            tags:
                - tasks
    /healthcheck:
        get:
            description: Checks if the application is running.
            operationId: healthcheckEndpoint
            responses:
                "200":
                    $ref: '#/responses/healthCheckResponse'
            summary: Health check endpoint.
            tags:
                - healthcheck
    /tasks:
        get:
            description: Fetches all tasks from the database.
            operationId: getAllTasksEndpoint
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/allTasksResponse'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Get all tasks.
            tags:
                - tasks
        post:
            consumes:
                - application/json
            description: Inserts a new task and its items into the database.
            operationId: createTaskEndpoint
            produces:
                - application/json
            responses:
                "201":
                    $ref: '#/responses/taskCreatedResponse'
                "400":
                    $ref: '#/responses/badRequestError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Create a new task.
            tags:
                - tasks
    /tasks/{id}:
        delete:
            description: Removes a task from the database based on its ID.
            operationId: deleteTaskEndpoint
            parameters:
                - description: The ID of the task to retrieve.
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/successfullyDeletedResponse'
                "400":
                    $ref: '#/responses/invalidTaskIdError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Delete a task by ID.
            tags:
                - tasks
        get:
            description: Fetches a task by its ID from the database.
            operationId: getTaskEndpoint
            parameters:
                - description: The ID of the task to retrieve.
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/taskResponse'
                "400":
                    $ref: '#/responses/invalidTaskIdError'
                "404":
                    $ref: '#/responses/notFoundError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Get a task by ID.
            tags:
                - tasks
        put:
            description: Updates a task by its ID with the details provided in the request body.
            operationId: updateTaskEndpoint
            parameters:
                - description: The ID of the task to update.
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
                - description: The details of the task to update.
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/Task'
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/taskResponse'
                "400":
                    $ref: '#/responses/badRequestError'
                "404":
                    $ref: '#/responses/notFoundError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Update an existing task.
            tags:
                - tasks
    /tasks/{taskID}/assign/{userID}:
        patch:
            description: Assigns a user to a specific task based on task and user IDs.
            operationId: assignTaskEndpoint
            parameters:
                - description: The ID of the task to be assigned.
                  format: int64
                  in: path
                  name: taskID
                  required: true
                  type: integer
                  x-go-name: TaskID
                - description: The ID of the user to which the task will be assigned.
                  format: int64
                  in: path
                  name: userID
                  required: true
                  type: integer
                  x-go-name: UserID
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/taskResponse'
                "400":
                    $ref: '#/responses/invalidIdError'
                "404":
                    $ref: '#/responses/notFoundError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Assign a user to a task.
            tags:
                - tasks
    /users/{userID}/tasks/assigned:
        get:
            description: Returns a list of tasks assigned to a user based on their ID.
            operationId: getUserAssignedTasksEndpoint
            parameters:
                - description: The ID of the user whose assigned tasks are to be retrieved.
                  format: int64
                  in: path
                  name: userID
                  required: true
                  type: integer
                  x-go-name: UserID
            produces:
                - application/json
            responses:
                "200":
                    $ref: '#/responses/allTasksResponse'
                "400":
                    $ref: '#/responses/invalidIdError'
                "500":
                    $ref: '#/responses/internalServerError'
            schemes:
                - http
                - https
            summary: Get tasks assigned to a specific user.
            tags:
                - tasks
produces:
    - application/json
responses:
    allCommentsResponse:
        description: Response for getting all comments for a task
        schema:
            items:
                $ref: '#/definitions/TaskComment'
            type: array
    allTasksResponse:
        description: Response for getting all tasks
        schema:
            items:
                $ref: '#/definitions/Task'
            type: array
    badRequestError:
        description: Bad request due to client-side error, e.g., invalid request body.
        headers:
            error:
                type: string
    healthCheckResponse:
        description: Health check response indicating the status and details of the service.
        headers:
            environment:
                type: string
            status:
                type: string
            version:
                type: string
    internalServerError:
        description: Server encountered a problem.
        headers:
            error:
                type: string
    invalidIdError:
        description: Invalid ID error due to client-side error, specifically when trying to assign a user to a task.
        headers:
            error:
                type: string
    invalidTaskIdError:
        description: Bad request due to client-side error, e.g., invalid task ID.
        headers:
            error:
                type: string
    notFoundError:
        description: Indicates the task was not found.
        headers:
            error:
                type: string
    successfullyDeletedResponse:
        description: Response indicating successful deletion.
        headers:
            message:
                type: string
    taskCommentCreatedResponse:
        description: Response for a successfully created task comment.
        schema:
            $ref: '#/definitions/TaskComment'
    taskCreatedResponse:
        description: Response for successfully created task.
        schema:
            $ref: '#/definitions/Task'
    taskResponse:
        description: Response for successfully retrieved task by ID.
        schema:
            $ref: '#/definitions/Task'
schemes:
    - http
swagger: "2.0"